- Refactor Location of PersonalizedExtraTraningExam.kt
