+++ app/src/main/java/de/wolfwarrior/thwtheorie/logik/personalizedRound/PersonalizedExtraTraningExam.kt
@@ -1,10 +1,12 @@
-package de.wolfwarrior.thwtheorie.logik
+package de.wolfwarrior.thwtheorie.logik.personalizedRound
 
 import Question
+import de.wolfwarrior.thwtheorie.logik.GeneralImplementationK
+import de.wolfwarrior.thwtheorie.logik.TheorieLogikInterface
 import java.util.Random
 
 
-class PersonalizedExtraTraningExam :GeneralImplementationK(), TheorieLogikInterface{
+class PersonalizedExtraTraningExam : GeneralImplementationK(), TheorieLogikInterface {
     override fun getResults(): String {
         val questionCounts = wrong + right
         return "Gratulation du den Test erfolgreich durchlaufen. Dabei hast du $questionCounts Fragen beantwortet, davon waren $right richtig und $wrong falsch."
